# slice_pushdown projections