closes #3135

A slow development build show these results

```python
n = int(1e6)
df = pl.DataFrame([
    pl.arange(0, n, eager=True).alias("a"),
    pl.arange(0, n, eager=True).alias("b"),
])
```

```python
%%time

df.lazy().select([
    pl.all().cast(str).suffix("foo"),
    (pl.col("a") * 10).alias("binary"),
    pl.when(pl.col("a") > 10).then(1).otherwise(2).alias("ternary")
]).limit(10).collect()

CPU times: user 7.94 ms, sys: 3.26 ms, total: 11.2 ms
Wall time: 2.32 ms
```

## without optimization
```python
%%time

df.lazy().select([
    pl.all().cast(str).suffix("foo"),
    (pl.col("a") * 10).alias("binary"),
    pl.when(pl.col("a") > 10).then(1).otherwise(2).alias("ternary")
]).limit(10).collect(slice_pushdown=False)

CPU times: user 833 ms, sys: 0 ns, total: 833 ms
Wall time: 339 ms

```

## Optimized logical plan

```
SELECT 4 COLUMNS: [col("a").slice(offset=0i64, length=10u32).cast(Utf8).alias("afoo"), col("b").slice(offset=0i64, length=10u32).cast(Utf8).alias("bfoo"), [(col("a").slice(offset=0i64, length=10u32)) * (10i64)].alias("binary"), 
WHEN [(col("a").slice(offset=0i64, length=10u32)) > (10i64)]
	1i32
OTHERWISE
	2i32.alias("ternary")]
FROM
DATAFRAME(in-memory): ["a", "b"];
	project 2/2 columns	|	details: Some([col("a"), col("b")]);
	selection: "None"



```